renamed front end files and reconnected routes. All components are being exported from the components index.js, so can import now with destructuring from index.js
