+++ client/components/allProducts.spec.js
