+++ client/components/userHome.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import {connect} from 'react-redux'
-import {Cart} from './cart'
+// import {Cart} from './cart'
 // import {AllProducts} from './AllProducts'
 
 /**
